refactor: move submit handler to output module
