+++ Project-Root/process.js
@@ -7,45 +7,8 @@
 
 // Nhập dữ liệu cấu hình và state
 import { jarsConfig, state, updateState } from './data.js';
-import { renderOutput, hideModal } from './output.js';
 
 // Export các hàm để các module khác có thể dùng
-export function handleTransactionSubmit(e) {
-    e.preventDefault();
-
-    const type = document.getElementById('transaction-type').value;
-    const amount = parseFloat(document.getElementById('transaction-amount').value);
-    const description = document.getElementById('transaction-description').value;
-    const jarId = document.getElementById('transaction-jar').value;
-
-    if (isNaN(amount) || amount <= 0 || !description || !jarId) {
-        console.error("Dữ liệu giao dịch không hợp lệ");
-        return;
-    }
-
-    const currentBalance = state.jars[jarId]?.balance || 0;
-    if (type === 'expense' && currentBalance < amount) {
-        console.error(`Số dư trong hũ "${jarsConfig[jarId].name}" không đủ.`);
-        return;
-    }
-
-    state.jars[jarId].balance += (type === 'income' ? amount : -amount);
-
-    const newTransaction = {
-        id: Date.now().toString(),
-        type,
-        amount,
-        description,
-        jarId,
-        createdAt: new Date().toISOString()
-    };
-    state.transactions.unshift(newTransaction);
-
-    saveStateToLocalStorage();
-    renderOutput();
-    hideModal('transaction-modal');
-    document.getElementById('transaction-form').reset();
-}
 
 export function calculateTotalBalance() {
     state.totalBalance = Object.values(state.jars).reduce((sum, jar) => sum + (jar.balance || 0), 0);
